Fix encoding issues when using Copy button
Previously the Copy button was using the GitHub API to fetch the file
contents in base64 form, and then decoding that. This caused unicode
symbols to be incorrectly encoded (such as the copyright symbol in
several files). This commit changes the Copy button to instead fetch the
file directly, which fixes the encoding issue.